# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Delivery API",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insira o token JWT no formato: Bearer {token}"
    }
  },
...
  "paths": {
...
    "/reviews": {
      "post": {
        "tags": [
          "Avaliações"
        ],
        "summary": "Enviar uma avaliação após entrega",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "review",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "orderId": {
                  "type": "string"
                },
                "companyId": {
                  "type": "integer"
                },
                "rating": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 5
                },
                "comment": {
                  "type": "string"
                }
              },
              "required": [
                "orderId",
                "companyId",
                "rating"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Avaliação registrada com sucesso"
          },
          "400": {
            "description": "Erro de validação ou já avaliado"
          }
        }
      }
    },
...
    "/coupons": {
      "get": {
        "tags": [
          "Cupons"
        ],
        "summary": "Listar todos os cupons do sistema",
        "responses": {
          "200": {
            "description": "Lista de cupons"
          }
        }
      },
      "post": {
        "tags": [
          "Cupons"
        ],
        "summary": "Criar novo cupom do sistema",
        "parameters": [
          {
            "in": "body",
            "name": "cupom",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "discount": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "minValue": {
                  "type": "number"
                },
                "isActive": {
                  "type": "boolean"
                }
              },
              "required": [
                "code",
                "discount",
                "description",
                "minValue",
                "isActive"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Cupom criado com sucesso"
          },
          "409": {
            "description": "Cupom com mesmo código já existe"
          }
        }
      }
    },
    "/coupons/{id}": {
      "patch": {
        "tags": [
          "Cupons"
        ],
        "summary": "Atualizar cupom por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "dados",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "discount": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "minValue": {
                  "type": "number"
                },
                "isActive": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cupom atualizado"
          },
          "404": {
            "description": "Cupom não encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Cupons"
        ],
        "summary": "Excluir cupom por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Cupom removido com sucesso"
          },
          "404": {
            "description": "Cupom não encontrado"
          }
        }
      }
    },
    "/user-coupons": {
      "get": {
        "tags": [
          "Cupons do Usuário"
        ],
        "summary": "Listar cupons do usuário logado",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de cupons"
          },
          "401": {
            "description": "Token não enviado"
          },
          "403": {
            "description": "Token inválido"
          }
        }
      }
    },
    "/user-coupons/{couponId}": {
      "post": {
        "tags": [
          "Cupons do Usuário"
        ],
        "summary": "Vincular cupom do sistema ao usuário",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "couponId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Cupom vinculado com sucesso"
          },
          "404": {
            "description": "Cupom não encontrado"
          },
          "409": {
            "description": "Cupom já vinculado"
          }
        }
      }
    },
    "/user-coupons/{id}": {
      "patch": {
        "tags": [
          "Cupons do Usuário"
        ],
        "summary": "Atualizar cupom do usuário (ex: usesLeft)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "dados",
            "schema": {
              "type": "object",
              "properties": {
                "usesLeft": {
                  "type": "integer"
                },
                "purchased": {
                  "type": "boolean"
                },
                "isActive": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cupom atualizado"
          },
          "404": {
            "description": "Cupom não encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Cupons do Usuário"
        ],
        "summary": "Remover cupom do usuário",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Cupom removido com sucesso"
          },
          "404": {
            "description": "Cupom não encontrado"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "Usuários"
        ],
        "summary": "Listar todos os usuários",
        "responses": {
          "200": {
            "description": "Lista de usuários"
          }
        }
      },
      "post": {
        "tags": [
          "Usuários"
        ],
        "summary": "Criar novo usuário",
        "parameters": [
          {
            "in": "body",
            "name": "usuario",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "senha": {
                  "type": "string"
                },
                "endereco": {
                  "type": "string"
                },
                "documento": {
                  "type": "string"
                },
                "celular": {
                  "type": "string"
                },
                "tipo": {
                  "type": "string",
                  "enum": [
                    "cliente",
                    "empresa"
                  ]
                },
                "codigoIndicadoPor": {
                  "type": "string",
                  "description": "Código de convite de quem indicou (opcional)"
                }
              },
              "required": [
                "nome",
                "email",
                "senha",
                "endereco",
                "documento",
                "celular",
                "tipo"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number"
                },
                "nome": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "tipo": {
                  "type": "string"
                },
                "codigoConvite": {
                  "type": "string",
                  "description": "Código exclusivo do usuário para compartilhar com amigos"
                }
              }
            }
          }
        }
      }
    },
    "/user-reviews": {
      "post": {
        "tags": [
          "Usuários"
        ],
        "summary": "Empresa avalia o usuário após pedido entregue",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "dados",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "orderId": {
                  "type": "string"
                },
                "rating": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 5
                }
              },
              "required": [
                "orderId",
                "rating"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuário avaliado com sucesso"
          },
          "400": {
            "description": "Erro de validação"
          },
          "404": {
            "description": "Pedido ou usuário não encontrado"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Usuários"
        ],
        "summary": "Buscar usuário por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      },
      "patch": {
        "tags": [
          "Usuários"
        ],
        "summary": "Atualizar usuário por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "dados",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "senha": {
                  "type": "string"
                },
                "endereco": {
                  "type": "string"
                },
                "documento": {
                  "type": "string"
                },
                "celular": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário atualizado"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Usuários"
        ],
        "summary": "Remover usuário por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário removido com sucesso"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      }
    },
...
  }
}